refactor(__test__)목데이터 분리
